src/main/java/com/bajiru/bank/service/impl/ProductCategoryServiceImpl.java
package com.bajiru.bank.service.impl;

import com.bajiru.bank.domain.ProductCategory;
import com.bajiru.bank.service.ProductCategoryService;
import com.bajiru.bank.service.repository.ProductCategoryRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * @Description create product category service implments
 * @AuthorName StevenWu
 * @CreateDateTime 2022-01-16-7:47 AM
 */
@Service
public class ProductCategoryServiceImpl implements ProductCategoryService {

    @Autowired
    private ProductCategoryRepository repository;


    @Override
    public ProductCategory findOne(Integer categoryId) {
        return repository.findById(categoryId).orElse(null);
    }

    @Override
    public List<ProductCategory> findAll() {
        return repository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return repository.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return repository.save(productCategory);
    }
}
